Format slack message and create const for scores
