+++ lighthouse/jest.config.lighthouse.js
