+++ package.json
@@ -8,8 +8,8 @@
     "start": "ng serve",
     "build": "ng build",
     "test:unit": "jest --maxWorkers=2 --verbose  --coverage",
-    "test:lighthouse": "jest -c jest.config.lighthouse.js",
-    "test:lighthouse:debug": "node --inspect node_modules/.bin/jest --watch -c jest.config.lighthouse.js --runInBand",
+    "test:lighthouse": "jest -c ./lighthouse/jest.config.lighthouse.js",
+    "test:lighthouse:debug": "node --inspect node_modules/.bin/jest --watch -c ./lighthouse/jest.config.lighthouse.js --runInBand",
     "lint": "ng lint"
   },
   "dependencies": {
